requirements.txt
open3d
ursina
panda3d
openai
python-dotenv